Move Opcode to primitives
With all of the prior adjustments to the type, Opcode can now be moved
to primitives.

Move Opcode to primitives and re-export from bitcoin.
